move ConvertSingleOpToCommunication to a new separate file
