+++ csrc/host_ir/lower_to_communication.h
@@ -1,0 +1,20 @@
+// clang-format off
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2024-present NVIDIA CORPORATION & AFFILIATES.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+// clang-format on
+#pragma once
+
+#include <host_ir/lower.h>
+#include <ir/all_nodes.h>
+
+namespace nvfuser {
+
+std::vector<Expr*> convertSingleOpToCommunication(
+    Expr* c,
+    DeviceIdxType my_device_idx,
+    const HostIrLowerParams& params);
+
+} // namespace nvfuser
